+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/HttpTableLookupFunction.java
@@ -1,5 +1,7 @@
 package com.getindata.connectors.http.internal.table.lookup;
 
+import java.util.Collection;
+import java.util.Collections;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -8,17 +10,16 @@ import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.api.common.serialization.DeserializationSchema;
-import org.apache.flink.table.data.GenericRowData;
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.functions.FunctionContext;
-import org.apache.flink.table.functions.TableFunction;
+import org.apache.flink.table.functions.LookupFunction;
 
 import com.getindata.connectors.http.internal.PollingClient;
 import com.getindata.connectors.http.internal.PollingClientFactory;
 import com.getindata.connectors.http.internal.utils.SerializationSchemaUtils;
 
 @Slf4j
-public class HttpTableLookupFunction extends TableFunction<RowData> {
+public class HttpTableLookupFunction extends LookupFunction {
 
     private final PollingClientFactory<RowData> pollingClientFactory;
 
@@ -50,32 +51,22 @@ public class HttpTableLookupFunction extends TableFunction<RowData> {
 
     @Override
     public void open(FunctionContext context) throws Exception {
-        super.open(context);
-
         this.responseSchemaDecoder.open(
             SerializationSchemaUtils
                 .createDeserializationInitContext(HttpTableLookupFunction.class));
 
         this.localHttpCallCounter = new AtomicInteger(0);
         this.client = pollingClientFactory
-            .createPollClient(options, responseSchemaDecoder);
+                .createPollClient(options, responseSchemaDecoder);
 
-        context
-            .getMetricGroup()
-            .gauge("http-table-lookup-call-counter", () -> localHttpCallCounter.intValue());
+        context.getMetricGroup()
+                .gauge("http-table-lookup-call-counter", () -> localHttpCallCounter.intValue());
     }
 
-    /**
-     * This is a lookup method which is called by Flink framework in a runtime.
-     */
-    public void eval(Object... keys) {
-        lookupByKeys(keys)
-            .ifPresent(this::collect);
-    }
-
-    public Optional<RowData> lookupByKeys(Object[] keys) {
-        RowData rowData = GenericRowData.of(keys);
+    @Override
+    public Collection<RowData> lookup(RowData keyRow) {
         localHttpCallCounter.incrementAndGet();
-        return client.pull(rowData);
+        Optional<RowData> result = client.pull(keyRow);
+        return result.map(Collections::singletonList).orElse(Collections.emptyList());
     }
 }